refactor: stories -> drops
