+++ src/Routes.tsx
@@ -1,9 +1,9 @@
+import { StoriesRoutes, StoriesRoutesConfig } from 'modules/drops/Routes';
 import { MarketRoutes, MarketRoutesConfig } from 'modules/market/Routes';
 import {
   ProfileRoutes,
   ProfileRoutesConfig,
 } from 'modules/profile/ProfileRoutes';
-import { StoriesRoutes, StoriesRoutesConfig } from 'modules/stories/Routes';
 import { Route, Switch } from 'react-router-dom';
 import { BrandRoutes, BrandRoutesConfig } from './modules/brand/BrandRoutes';
 import {
