+++ src/modules/drops/components/StoriesContainer/StoriesContainer.tsx
