+++ src/modules/drops/components/StoriesContainer/useStoriesContainerStyles.ts
