+++ src/modules/drops/components/StoriesOwner/assets/VerifiedIcon.tsx
