+++ src/modules/drops/screens/DropDetails/index.ts
