+++ src/modules/drops/screens/Stories/components/Drop/useDropStyles.ts
