+++ src/modules/drops/screens/Stories/components/DropsSection/DropsSection.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography } from '@material-ui/core';
-import { StoriesContainer } from 'modules/stories/components/StoriesContainer';
-import { StoriesOwner } from 'modules/stories/components/StoriesOwner';
-import { StoriesRoutesConfig } from 'modules/stories/Routes';
+import { StoriesContainer } from 'modules/drops/components/StoriesContainer';
+import { StoriesOwner } from 'modules/drops/components/StoriesOwner';
+import { StoriesRoutesConfig } from 'modules/drops/Routes';
 import { Section } from 'modules/uiKit/Section';
 import React, { ChangeEvent, useCallback, useEffect, useState } from 'react';
 import { uid } from 'react-uid';
