+++ src/modules/drops/screens/Stories/components/DropsTabs/index.ts
