+++ src/modules/drops/screens/Stories/components/StoriesChip/StoriesChip.tsx
