+++ src/modules/drops/screens/Stories/components/StoriesChip/assets/LightningIcon.tsx
