+++ src/modules/drops/screens/Stories/components/StoriesSlider/index.ts
