+++ src/modules/drops/screens/Stories/components/StoriesSlider/useStoriesSliderStyles.ts
