+++ src/modules/drops/screens/Stories/components/StoriesSliderItem/StoriesSliderItem.tsx
