+++ src/modules/drops/screens/Stories/components/StoriesSliderItem/StoriesSliderItemSkeleton.tsx
