+++ src/modules/drops/screens/Stories/components/StoriesSliderItem/index.ts
