+++ src/modules/drops/screens/Stories/components/StoriesSliderItem/useStoriesSliderItemStyles.ts
